[Chore #164185920] Refactor v2 file names
