+++ api/__init__.py
@@ -10,12 +10,12 @@ from .v1.views.party_view import party_api
 from .v1.views.office_view import office_api
 from instance.config import application_config
 from .db_conn import create_tables, drop_tables
-from .v2.views.user_view import user_api_v2
-from .v2.views.office_view import office_api_v2
-from .v2.views.parties_view import parties_api_v2
-from .v2.views.candidate_view import candidate_api_v2
-from .v2.views.votes_view import votes_api_v2
-from .v2.views.results_view import result_api_v2
+from .v2.views.user import user_api_v2
+from .v2.views.office import office_api_v2
+from .v2.views.parties import parties_api_v2
+from .v2.views.candidate import candidate_api_v2
+from .v2.views.votes import votes_api_v2
+from .v2.views.results import result_api_v2
 
 
 # Error Handler Methods
