+++ api/v2/views/votes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request, make_response, jsonify
-from api.v2.models.votes_model import VoteModel
+from api.v2.models.votes import VoteModel
 from flask_jwt_extended import jwt_required
 
 votes_api_v2 = Blueprint('votes_v2', __name__, url_prefix="/api/v2")
