+++ tests/v2tests/__init__.py
@@ -2,12 +2,12 @@ import unittest
 # Imports create app function to set testing config
 from run import create_app
 # from api.db_conn import execute_drop_queries
-from api.v2.models.user_model import UserModelDb
-from api.v2.models.office_model import OfficesModelDb
-from api.v2.models.candidate_model import CandidateModel
-from api.v2.models.votes_model import VoteModel
+from api.v2.models.user import UserModelDb
+from api.v2.models.office import OfficesModelDb
+from api.v2.models.candidate import CandidateModel
+from api.v2.models.votes import VoteModel
 from api.db_conn import create_tables, drop_tables, close_connection
-from api.v2.models.parties_model import PartiesModelDb
+from api.v2.models.parties import PartiesModelDb
 import json
